Restructuring Homepage and HelpListView components
