+++ frontend/src/pages/HelpListView.jsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from "react";
 import { useSearchParams } from "react-router-dom";
+import { DatePicker, Space } from "antd";
 import HelpList from "../components/HelpList";
 import Loading from "../presentations/Loading";
 import getHelpRequests from "../api";
@@ -8,16 +9,19 @@ function HelpListView() {
   const [helpList, setHelpList] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
 
-  const [dateParams, setDateParams] = useSearchParams();
-  let dateQuery = dateParams.get("date");
+  // const [dateParams, setDateParams] = useSearchParams();
+  // let dateQuery = dateParams.get("date");
   const [typeParams, setTypeParams] = useSearchParams();
   let typeQuery = typeParams.get("type");
 
-  function handleDateChange(event) {
-    const newParams = new URLSearchParams(dateParams);
-    newParams.set("date", event.target.value);
-    setDateParams(newParams);
-    dateQuery = newParams.get("date");
+  function handleFromDateChange(date, dateString) {
+    console.log("date:", date, "dateString:", dateString);
+    fromDate = dateString;
+  }
+
+  function handleEndDateChange(date, dateString) {
+    console.log("date:", date, "dateString:", dateString);
+    endDate = dateString;
   }
 
   function handleTypeChange(event) {
@@ -29,35 +33,47 @@ function HelpListView() {
 
   useEffect(() => {
     let endpoint = "/api/help-requests";
-    if (dateQuery) {
-      endpoint += `?date=${dateQuery}`;
-    } else if (typeQuery) {
+    // if (dateQuery) {
+    //   endpoint += `?date=${dateQuery}`;
+    // } else
+    if (typeQuery) {
       endpoint += `?type=${typeQuery}`;
     }
-    getHelpRequests(path)
+    getHelpRequests(endpoint)
       .then((response) => {
+        // console.log(response.data, "data 1");
+        console.log(response.data.helpRequestsData, "data 2");
         setHelpList(response.data.helpRequestsData);
         setIsLoading(false);
       })
       .catch((err) => {
         console.log(err);
       });
-  }, [dateQuery, typeQuery]);
+  }, [
+    // dateQuery,
+    typeQuery,
+  ]);
+
+  // const newHelpList = helpList.filter((helpRequest) => {
+  //   return helpRequest.req_date > fromDate && helpRequest.req_date < endDate;
+  // });
 
   if (isLoading) {
     return <Loading text={"help requests near you"} />;
   }
   return (
     <>
-      <h2>Help Requests Near You</h2>
-      <label htmlFor="date">Filter requests by date needed: </label>
-      <select name="date-options" id="date" onChange={handleDateChange}>
-        <option value="">All Help Requests</option>
-        <option value="oneday">Needed today</option>
-        <option value="threedays">Needed in the next 3 days</option>
-        <option value="oneweek">Needed within a week</option>
-        <option value="onemonth">Needed within a month</option>
-      </select>
+      <p>Filter</p>
+      <label htmlFor="from-date">From: </label>
+      <Space id="from-date" direction="vertical">
+        <DatePicker onChange={handleFromDateChange} />
+      </Space>
+      <br />
+      <label htmlFor="end-date">To: </label>
+      <Space id="end-date" direction="vertical">
+        <DatePicker onChange={handleEndDateChange} />
+      </Space>
+      <br />
       <br />
       <label htmlFor="types">Filter requests by type: </label>
       <select name="type-options" id="types" onChange={handleTypeChange}>
@@ -68,9 +84,7 @@ function HelpListView() {
         <option value="vehicle">Vehicle</option>
         <option value="community">Community</option>
       </select>
-      <br />
       <HelpList helpList={helpList} />
-      <br />
     </>
   );
 }
